requirements.txt
addict==2.4.0
antlr4-python3-runtime==4.9.3
certifi==2022.12.7
charset-normalizer==2.1.1
cmake==3.25.0
contourpy==1.1.0
cycler==0.11.0
filelock==3.9.0
fonttools==4.42.0
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.4.8
importlib-metadata==6.8.0
importlib-resources==6.0.0
Jinja2==3.1.2
kiwisolver==1.4.4
kornia==0.7.0
lit==15.0.7
lpips==0.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.7.2
mdurl==0.1.2
mmcv==1.6.0
mmengine==0.8.4
mpmath==1.2.1
networkx==3.0
numpy==1.24.1
omegaconf==2.3.0
opencv-python==4.8.0.74
packaging==23.1
Pillow==9.3.0
platformdirs==3.10.0
psutil==5.9.5
Pygments==2.16.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-msssim==1.0.0
PyYAML==6.0.1
requests==2.28.1
rich==13.5.2
scipy==1.11.1
six==1.16.0
sympy==1.11.1
termcolor==2.3.0
tomli==2.0.1
torch==2.0.1+cu117
torch-scatter==2.1.1
torchaudio==2.0.2+cu117
torchvision==0.15.2+cu117
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.4.0
urllib3==1.26.13
yapf==0.40.1
zipp==3.16.2
